Bug Fixes + Removed Unused local
Fixed bug with the legendary animal, and its helper animals not deleting when you died. Was caused by incorrect variable names. Also in the process of fixing that realized that the delete secondary animal part only deleted one of the secondary animals. Secondary animal spawns are now stored in a local table, so that the delete ped can remove all the secondary animals that spawned not just one.